# Small speedup on autosizing column width