@Bykiev 2023-09-05
@tonyqus, we can also improve the speed if we will use parallel programming while looping over the rows. What do you think about it?