@Bykiev 2023-09-06
> parallel programming is somewhat dangerous to cause CPU 100% if you don't limit the parallel number.
> 
> We can try this. But we may need to carefully test this.

Thanks, I think these experiments for another PR, this one can be merged without any issues