@Bykiev 2023-09-06
Thought after applying caching this PR doesn't improve performance, the code is cleaner by removing unused code.

Benchmarks.

**Before:**
|         Method | RowCount | ColumnCount |     Mean |   Error |  StdDev |       Gen0 | Allocated |
|--------------- |--------- |------------ |---------:|--------:|--------:|-----------:|----------:|
| AutoSizeColumn |     1000 |           5 | 433.1 ms | 6.87 ms | 6.75 ms | 75000.0000 | 117.23 MB |


**After:**
|         Method | RowCount | ColumnCount |     Mean |   Error |  StdDev |       Gen0 | Allocated |
|--------------- |--------- |------------ |---------:|--------:|--------:|-----------:|----------:|
| AutoSizeColumn |     1000 |           5 | 428.3 ms | 6.04 ms | 5.65 ms | 75000.0000 | 117.23 MB |